created game stats component, modularized in game stats component, added loading dots component, and created custom hook for captured pieces
