+++ src/components/Dashboard/InGameStats/InGameStats.css
@@ -1,0 +1,145 @@
+
+.game-card{
+    background-color: #181818;
+    border-radius: 20px;
+
+    height: 100%;
+    width: 30vw;
+    margin-left: 10vw;
+}
+
+.game-info{
+    background-color: #3D3D3D;
+    border-radius: 20px;
+    margin-left: 5%;
+    height: 20%;
+    width: 90%;
+    margin-top: 5%;
+    margin-bottom: 5%;
+    display: flex;
+    flex-direction: row;
+}
+
+.game-stats{
+    display: flex;
+    flex-direction: column;
+    width: 90%;
+    padding-left: 5%;
+    justify-content: center;
+    gap: 10%;
+    
+}
+
+.game-info h3{
+    font-size: 18px;
+    padding: 0;
+    margin: 0;
+}
+
+.game-row-1{
+    display: flex;
+    flex-direction: row;
+    color: #FFFFFF;
+    text-align: left;
+    justify-content: left;
+    align-items: center;
+    gap: 2%;
+    
+}
+
+.game-row-2{
+    display: flex;
+    flex-direction: row;
+    color: #FFFFFF;
+    text-align: left;
+    justify-content: left;
+    align-items: center;
+    gap: 5%;
+    
+}
+
+.game-icon{
+    height: 30px;
+    color: #FFCD05;
+}
+
+.game-pieces-captured{
+    height: 15px;
+    flex: 1;
+}
+
+.game-pieces-captured-hidden{
+    visibility: hidden;
+}
+
+.game-row-3{
+    display: flex;
+    flex-direction: row;
+    text-align: left;
+    justify-content: left;
+    align-items: center;
+    gap: 5%;   
+}
+
+.player-turn{
+    background-color: #FFCD05;
+}
+
+.moves-header{
+    height: 6%;
+    text-align: center;
+    justify-content: center;
+    align-items: center;
+}
+
+.moves-header-text{
+    font-size: 18px;
+    height: 50%;
+    padding: 0;
+    margin: 0;
+    color: #FFE769;
+}
+
+.moves-icon{
+    padding: 0;
+    margin: 0;
+    height: 50%;
+}
+
+.moves-list{
+    text-align: left;
+    height: 27vh;
+    overflow: auto;
+    padding-left: 5%;
+    padding-right: 5%;
+}
+
+.moves-list li{
+    list-style: none;
+    text-decoration: none;
+    list-style-type: none;
+    
+}
+
+.moves-line{
+    color: #FFFFFF;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    
+    width: 100%;
+    height: 4.5vh;
+}
+
+.moves-scroll{
+    width: 105%;
+}
+
+.moves-line h3{
+    font-size: 14px;
+    flex: 1;
+}
+
+.move-number{
+    padding-left: 15%;
+}
